--- DiscordBot/launchwith.txt
@@ -1,1 +1,0 @@
-./forever [filename].py
